criado regras de negócio com data limit de no máximo 3 meses,exclusão apenas se não tiver crédito aprovado,alterado retorno de CustomerController para Classe CustomerView

+++ Bootcamp TQI Kotlin - Backend Developer/api credit system/credit-aplication-sytem/src/main/kotlin/com/me/group/credit/sytem/service/serviceImpl/CustomerServiceImpl.kt
@@ -1,7 +1,9 @@
 package com.me.group.credit.sytem.service.serviceImpl
 
 import com.me.group.credit.sytem.entity.Customer
+import com.me.group.credit.sytem.enums.Status
 import com.me.group.credit.sytem.exeception.BusinessException
+import com.me.group.credit.sytem.repository.CreditRepository
 import com.me.group.credit.sytem.repository.CustomerRepository
 import com.me.group.credit.sytem.service.ICustomerService
 import org.springframework.stereotype.Service
@@ -9,7 +11,8 @@ import java.lang.RuntimeException
 
 @Service
 class CustomerServiceImpl(
-        private val custumerRepository: CustomerRepository
+        private val custumerRepository: CustomerRepository,
+        private val creditRepository: CreditRepository
 ):ICustomerService {
 
     override fun save(customer: Customer): Customer {
@@ -24,10 +27,15 @@ class CustomerServiceImpl(
 
     override fun delete(idCostumer: Long): Boolean{
             val customer = findById(idCostumer)
-          if (customer !=null){
+             val lisCredits =  creditRepository.findAllByCustomer(idCostumer).filter {
+                    it.status == Status.APPROVED
+             }
+
+          if (customer !=null && lisCredits.isEmpty()){
               custumerRepository.delete(customer)
               return true
           }
+        throw BusinessException("Customer ${customer.fistName} have pending Crédit")
         return false
 
     }
